requirements.txt
# Core Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.11.1
pydantic-settings==2.8.1

# Telegram Integration
telethon==1.39.0
cryptg==0.5.0  # Speed up Telethon

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis Cache
redis==5.0.1
hiredis==2.3.2  # Speed up Redis

# Vector Database
qdrant-client==1.7.0

# AI/ML
openai==1.8.0
anthropic==0.8.1
tiktoken==0.5.2  # Token counting
langchain==0.1.0
langchain-openai==0.0.3

# Message Queue
aiokafka==0.10.0
kafka-python==2.0.2

# Data Processing
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
scikit-learn==1.4.0
matplotlib==3.8.2

# Technical Analysis
ta-lib==0.6.7
yfinance==0.2.33

# NLP and Text Processing
spacy==3.7.2
nltk==3.8.1

# Image Processing
pillow==10.2.0
opencv-python==4.9.0.80

# OCR
google-cloud-vision==3.5.0
pytesseract==0.3.10
easyocr==1.7.0

# API and Web
httpx==0.26.0
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.9

# Monitoring and Logging
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-propagator-b3==1.22.0
structlog==24.1.0
python-json-logger==2.0.7
apscheduler==3.10.4

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
pytz==2023.3.post1
croniter==2.0.1
tenacity==8.2.3  # Retry logic
cachetools==5.3.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# System Monitoring
psutil==5.9.8

